+++ scrooge-generator/src/main/scala/com/twitter/scrooge/frontend/ThriftParser.scala
@@ -237,8 +237,7 @@ class ThriftParser(
           transformedReq,
           typeAnnotations,
           fieldAnnotations,
-          comm,
-          fieldAnnotations.keySet.exists(_.startsWith("validation."))
+          comm
         )
     }
   )