+++ scrooge-thrift-validation/src/main/scala/com/twitter/scrooge/thrift_validation/DefaultAnnotations.scala
@@ -132,4 +132,6 @@ object DefaultAnnotations {
     ),
     "validation.UUID" -> AnnotationMetaData(None, Set(classOf[String]), classOf[UUID])
   )
+
+  private[twitter] val keys = metadata.keySet
 }